% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nbreak.R
\name{nbreak}
\alias{nbreak}
\title{nbreak}
\usage{
nbreak(string, n, loop = TRUE)
}
\arguments{
\item{string}{a string.}

\item{n}{an integer (n > 0).}

\item{loop}{Logical value. By default, loop = TRUE. If FALSE, it will break only a once.}
}
\description{
Breaks a line from the nth word, from n to n.
}
